Commit Messages
Added the folder 7.Commit-Messages.
It contains the Usual Knowledge-Check.md and also a list of 7 rules to
write a good Commit Message.

I added it because I want to proceed with the curriculum, and I couldn't
have done that if I had stopped with the last lesson.

I am going to Commit this way from now on because it makes me look
very professional.